Migrated numerics to BigInt/BigDecimal
This commit migrates Value's numeric types to BigInt and BigDecimal. The
basic idea is that overflow errors aren't great in a shell environment,
and not really necessary.

The main immediate consequence is that new errors can occur when
serializing Nu values to other formats. You can see this in changes to
the various serialization formats (JSON, TOML, etc.). There's a new
`CoerceInto` trait that uses the `ToPrimitive` trait from `num_traits`
to attempt to coerce a `BigNum` or `BigDecimal` into a target type, and
produces a `RangeError` (kind of `ShellError`) if the coercion fails.

Another possible future consequence is that certain performance-critical
numeric operations might be too slow. If that happens, we can introduce
specialized numeric types to help improve the performance of those
situations, based on the real-world experience.

+++ src/plugins/str.rs
@@ -258,13 +258,13 @@ fn main() {
 
 #[cfg(test)]
 mod tests {
-
     use super::{Action, ReplaceAction, Str};
     use indexmap::IndexMap;
     use nu::{
         CallInfo, EvaluatedArgs, Plugin, Primitive, ReturnSuccess, SourceMap, Span, Tag, Tagged,
         TaggedDictBuilder, TaggedItem, Value,
     };
+    use num_bigint::BigInt;
 
     impl Str {
         fn replace_with(&mut self, value: &str) {
@@ -600,7 +600,7 @@ mod tests {
             ReturnSuccess::Value(Tagged {
                 item: Value::Primitive(Primitive::Int(i)),
                 ..
-            }) => assert_eq!(*i, 10),
+            }) => assert_eq!(*i, BigInt::from(10)),
             _ => {}
         }
     }
